+++ runtime/datatypes/integer.ts
@@ -16,7 +16,7 @@ export function $compare(
 	if(value2 instanceof Red.RawInteger || value2 instanceof Red.RawFloat || value2 instanceof Red.RawMoney) {
 		return cmp(value1.value, value2.value);
 	} else if(value2 instanceof Red.RawChar) {
-		return cmp(value1.value, value2.char.charCodeAt(0));
+		return cmp(value1.value, value2.char);
 	} else if(value2 instanceof Red.RawPercent) {
 		return cmp(value1.value, value2.value/100);
 	} else if(value2 instanceof Red.RawTime) {
@@ -61,7 +61,7 @@ export function $$add(
 	} else if(right instanceof Red.RawPercent) {
 		return new Red.RawFloat(left.value + right.value / 100);
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value + right.char.charCodeAt(0));
+		return new Red.RawInteger(left.value + right.char);
 	} else if(right instanceof Red.RawMoney) {
 		return new Red.RawMoney(+(left.value + right.value).toFixed(2));
 	} else if(right instanceof Red.RawPair) {
@@ -85,7 +85,7 @@ export function $$add(
 			return new Red.RawVector(
 				values
 					.slice(right.index-1)
-					.map(v => new Red.RawChar(String.fromCharCode(left.value + v.char.charCodeAt(0))))
+					.map(v => new Red.RawChar(left.value + v.char))
 			);
 		} else {
 			return new Red.RawVector(
@@ -111,7 +111,7 @@ export function $$subtract(
 	} else if(right instanceof Red.RawPercent) {
 		return new Red.RawFloat(left.value - right.value / 100);
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value - right.char.charCodeAt(0));
+		return new Red.RawInteger(left.value - right.char);
 	} else if(right instanceof Red.RawMoney) {
 		return new Red.RawMoney(+(left.value - right.value).toFixed(2));
 	} else if(right instanceof Red.RawPair) {
@@ -135,7 +135,7 @@ export function $$subtract(
 			return new Red.RawVector(
 				values
 					.slice(right.index-1)
-					.map(v => new Red.RawChar(String.fromCharCode(left.value - v.char.charCodeAt(0))))
+					.map(v => new Red.RawChar(left.value - v.char))
 			);
 		} else {
 			return new Red.RawVector(
@@ -161,7 +161,7 @@ export function $$multiply(
 	} else if(right instanceof Red.RawPercent) {
 		return new Red.RawFloat(left.value * (right.value / 100));
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value * right.char.charCodeAt(0));
+		return new Red.RawInteger(left.value * right.char);
 	} else if(right instanceof Red.RawMoney) {
 		return new Red.RawMoney(+(left.value * right.value).toFixed(2));
 	} else if(right instanceof Red.RawPair) {
@@ -185,7 +185,7 @@ export function $$multiply(
 			return new Red.RawVector(
 				values
 					.slice(right.index-1)
-					.map(v => new Red.RawChar(String.fromCharCode(left.value * v.char.charCodeAt(0))))
+					.map(v => new Red.RawChar(left.value * v.char))
 			);
 		} else {
 			return new Red.RawVector(
@@ -211,7 +211,7 @@ export function $$divide(
 	} else if(right instanceof Red.RawPercent) {
 		return new Red.RawFloat(left.value / (right.value / 100));
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(Math.floor(left.value / right.char.charCodeAt(0)));
+		return new Red.RawInteger(Math.floor(left.value / right.char));
 	} else if(right instanceof Red.RawMoney) {
 		return new Red.RawMoney(+(left.value / right.value).toFixed(2));
 	} else if(right instanceof Red.RawPair) {
@@ -235,7 +235,7 @@ export function $$divide(
 			return new Red.RawVector(
 				values
 					.slice(right.index-1)
-					.map(v => new Red.RawChar(String.fromCharCode(Math.floor(left.value / v.char.charCodeAt(0)))))
+					.map(v => new Red.RawChar(Math.floor(left.value / v.char)))
 			);
 		} else {
 			return new Red.RawVector(
@@ -261,7 +261,7 @@ export function $$remainder(
 	} else if(right instanceof Red.RawPercent) {
 		return new Red.RawFloat(left.value % (right.value / 100));
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value % right.char.charCodeAt(0));
+		return new Red.RawInteger(left.value % right.char);
 	} else if(right instanceof Red.RawMoney) {
 		return new Red.RawMoney(+(left.value % right.value).toFixed(2));
 	} else if(right instanceof Red.RawPair) {
@@ -283,7 +283,7 @@ export function $$remainder(
 			return new Red.RawVector(
 				values
 					.slice(right.index-1)
-					.map(v => new Red.RawChar(String.fromCharCode(left.value % v.char.charCodeAt(0))))
+					.map(v => new Red.RawChar(left.value % v.char))
 			);
 		} else {
 			return new Red.RawVector(
@@ -307,7 +307,7 @@ export function $$and_t(
 	if(right instanceof Red.RawInteger) {
 		return new Red.RawInteger(left.value & right.value);
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value & right.toNormalChar().charCodeAt(0));
+		return new Red.RawInteger(left.value & right.char);
 	} else if(right instanceof Red.RawPair) {
 		return new Red.RawPair(left.value & right.x, left.value & right.y);
 	} else if(right instanceof Red.RawVector) {
@@ -316,7 +316,7 @@ export function $$and_t(
 		if(Red.RawVector.isInteger(values)) {
 			return new Red.RawVector(values.map(int => new Red.RawInteger(left.value & int.value)));
 		} else if(Red.RawVector.isChar(values)) {
-			return new Red.RawVector(values.map(int => new Red.RawChar(String.fromCharCode(left.value & int.toNormalChar().charCodeAt(0)))));
+			return new Red.RawVector(values.map(ch => new Red.RawChar(left.value & ch.char)));
 		} else {
 			throw new Error("error!");
 		}
@@ -340,7 +340,7 @@ export function $$or_t(
 	if(right instanceof Red.RawInteger) {
 		return new Red.RawInteger(left.value | right.value);
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value | right.toNormalChar().charCodeAt(0));
+		return new Red.RawInteger(left.value | right.char);
 	} else if(right instanceof Red.RawPair) {
 		return new Red.RawPair(left.value | right.x, left.value | right.y);
 	} else if(right instanceof Red.RawVector) {
@@ -349,7 +349,7 @@ export function $$or_t(
 		if(Red.RawVector.isInteger(values)) {
 			return new Red.RawVector(values.map(int => new Red.RawInteger(left.value | int.value)));
 		} else if(Red.RawVector.isChar(values)) {
-			return new Red.RawVector(values.map(int => new Red.RawChar(String.fromCharCode(left.value | int.toNormalChar().charCodeAt(0)))));
+			return new Red.RawVector(values.map(ch => new Red.RawChar(left.value | ch.char)));
 		} else {
 			throw new Error("error!");
 		}
@@ -366,7 +366,7 @@ export function $$xor_t(
 	if(right instanceof Red.RawInteger) {
 		return new Red.RawInteger(left.value ^ right.value);
 	} else if(right instanceof Red.RawChar) {
-		return new Red.RawInteger(left.value ^ right.toNormalChar().charCodeAt(0));
+		return new Red.RawInteger(left.value ^ right.char);
 	} else if(right instanceof Red.RawPair) {
 		return new Red.RawPair(left.value ^ right.x, left.value ^ right.y);
 	} else if(right instanceof Red.RawVector) {
@@ -375,7 +375,7 @@ export function $$xor_t(
 		if(Red.RawVector.isInteger(values)) {
 			return new Red.RawVector(values.map(int => new Red.RawInteger(left.value ^ int.value)));
 		} else if(Red.RawVector.isChar(values)) {
-			return new Red.RawVector(values.map(int => new Red.RawChar(String.fromCharCode(left.value ^ int.toNormalChar().charCodeAt(0)))));
+			return new Red.RawVector(values.map(ch => new Red.RawChar(left.value ^ ch.char)));
 		} else {
 			throw new Error("error!");
 		}